backend/requirements.txt
# backend/requirements.txt
faker
PyYAML
python-dotenv
sqlalchemy
psycopg2-binary  # Opymysql si usas MySQL
pandas
flask 
flask-cors